+++ include/criteria.h
@@ -8,15 +8,6 @@
 
 struct mako_config;
 
-// State is intended to work as a bitmask, so if more need to be added in the
-// future, this should be taken into account.
-enum mako_parse_state {
-	MAKO_PARSE_STATE_NORMAL = 0,
-	MAKO_PARSE_STATE_ESCAPE = 1,
-	MAKO_PARSE_STATE_QUOTE = 2,
-	MAKO_PARSE_STATE_QUOTE_ESCAPE = 3,
-};
-
 // Stores whether or not each field was part of the criteria specification, so
 // that, for example, "not actionable" can be distinguished from "don't care".
 // This is unnecessary for string fields, but it's best to just keep it
